+++ src/functions/init.js
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) 2020. by Pablo Klaschka
+ */
+
+const lang = require('xd-localization-helper');
+const SelectionChecker = require('../helpers/check-selection');
+
+/**
+ * Initialize all necessary components and check the selection for compatibility
+ * @param {XDSelection} selection
+ * @return {Promise<boolean>} Selection contains text layer?
+ */
+module.exports = async function init(selection) {
+    await lang.load();
+    let checker = new SelectionChecker(selection);
+    return checker.oneOrMore('Text');
+};
