+++ src/functions/placeholder-text/apply-text.js
