+++ src/functions/settings/fetch-settings.js
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2020. by Pablo Klaschka
+ */
+
+const storage = require('xd-storage-helper');
+
+/**
+ * Fetch the settings from storage
+ * @returns {Promise<{text: string, terminationString: string, includeLineBreaks: boolean, trim: boolean}>}
+ */
+module.exports = async function fetchSettings() {
+    return await storage.get('loremOptions', {
+        text: 'lorem-lat',
+        terminationString: '.',
+        includeLineBreaks: true,
+        trim: false
+    });
+};
